added ma_editor to new site system
