+++ fe_head/ma_editor.head.php
@@ -1,0 +1,7 @@
+<?php require_once("auth.php"); ?>
+
+<script>
+    
+var const_NE_DIR_RELATIVE = "<?php echo NE_DIR_RELATIVE; ?>";
+   
+</script>
